[MASSEMBLY-675] Maven Assembly packaging wildcard-excluded dependencies
Since Maven 3 requirement, there is no need to resolve any dependencies
manually, and we can rely on requiresDependencyResolution. Initial patch
failed when module sets were involved: instead of getting the list of
transitive/direct dependencies on the current project, they need to be
fetched for each project so that it encompasses the possible modules in
module sets.

All the dead code as a result of this change is removed.

This closes #2.