MASSEMBLY-775 outputDirectory is completely irrelevant for absolute paths
The old code did log some warn based on an absolute path in the outputDirectory.
But that's actually irrelevant as this is always the directory inside the zip/tar, etc.
That means it is _always_ interpreted as relative!
The only possible problem is the 'directory' attribute which
denotes the source path!