Do not download whole universe
Problem: try to build maven master with empty repository,
you will notice that we get things from repository.sonatype.org
repository, that is wrong. Moreover, if you look carefully, you
will notice that we download/resolve totally unrelated things
to our build, like felix plugin etc. For fun, it downloads
Maven 2.2.0, Maven 2.0.7 etc as well.

Without PR:
https://gist.github.com/cstamas/badb32a25dbe444679774611f61d94b1

Reason: m-assembly-p "rebuilds" all the project of depSet to
get MavenProject instance of them and transitive deps, but alas,
it was doing it too eagerly: it was getting even plugins that
built our dependendencies, while we are really not want them.
So, instead to re-use the "project building request" of
currently built project, just create a simple PBR that is
tuned "just enough" to get results we want, nothing more.
We do not want the plugins that built our dependencies.

With PR:
https://gist.github.com/cstamas/1e9dfd0384dceda7a8195edfc1bfb61b