[MASSEMBLY-1008] Wrong use of dependency scopes
This IT demonstrates how assembly plugin (among others)
does wrongly, the code probably originates from Maven2
times.

What happens:
* "single" mojo resolves test project dependencies
* then it reads assembly descriptor
* then uses project dependencies to deliver descriptor contents

But, here is a problem that IT demonstrates: the delivered list
of files are NOT runtime scoped, they are "test scope filtered
for runtime leaves" which is not the same thing. By the way,
dependency plugin demonstrates same behaviour, wrongly assumes
that "test" graph is larger "runtime" tree, but that is not true
nor was never true since Maven3 (resolver), it was true ONLY in
Maven2 times.